Standardize component names and configure absolute imports
