+++ src/App.tsx
@@ -1,24 +1,24 @@
 import { Routes, Route } from 'react-router-dom'
-import { Header } from '@components'
+import { Header, Footer } from '@components'
 import { ROUTES } from '@constants'
-import { Authorization } from '@pages'
-import './App.css'
+import { Authorization, MainPage } from '@pages'
 
 function App() {
     return (
         <>
             <Header />
-                <Routes>
-                    <Route
-                        path={ROUTES.REGISTER}
-                        element={<div>Регистрация</div>}
-                    />
-                    <Route path={ROUTES.LOGIN} element={<Authorization />} />
-                    <Route path={ROUTES.NOT_FOUND} element={<div>Ошибка</div>} />
-                </Routes>
-            {/* <Footer /> */}
+            <Routes>
+                <Route path={ROUTES.MAIN_PAGE} element={<MainPage />} />
+                <Route
+                    path={ROUTES.REGISTER}
+                    element={<div>Регистрация</div>}
+                />
+                <Route path={ROUTES.LOGIN} element={<Authorization />} />
+                <Route path={ROUTES.NOT_FOUND} element={<div>Ошибка</div>} />
+            </Routes>
+            <Footer />
         </>
-    );
+    )
 }
 
 export default App;
