--- src/components/Footer.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Footer: React.FC = () => {
-  return (
-    <footer className="bg-blue-600 text-white p-4 text-center w-screen fixed bottom-0 left-0 right-0">
-      <p>Команда: Дарья Басынина, Алексей Осипенко, Алексей Пономарев, Дарья Слободская 2025</p>
-    </footer>
-  );
-};
-
-export { Footer };
