--- src/components/Header.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Header: React.FC = () => {
-  return (
-    <header>
-      <p>Заглушка</p>
-    </header>
-  );
-};
-
-export { Header };
