+++ src/components/book-card/book-card.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import { Button } from './Button';
+import { Button } from '@components'
 
 type BookCardProps = {
   id: number;
@@ -9,7 +8,7 @@ type BookCardProps = {
   image: string; 
 }
 
-const BookCard: React.FC<BookCardProps> = ({ price, title, author, image }) => {
+export const BookCard: React.FC<BookCardProps> = ({ price, title, author, image }) => {
   return (
     <div className="border border-gray-300 rounded p-2 text-center">
       <img src={image} alt={title} className="w-full h-60 object-cover mb-2" />
@@ -23,5 +22,3 @@ const BookCard: React.FC<BookCardProps> = ({ price, title, author, image }) => {
     </div>
   );
 };
-
-export  { BookCard };
