+++ src/components/buttons/form-button.jsx
@@ -1,4 +1,4 @@
-export const Button = ({ children, className, ...props }) => {
+export const FormButton = ({ children, className, ...props }) => {
 	return (
 		<button
 			className={`flex items-center justify-center h-10 text-base rounded-3xl pl-6 pr-6 ${className}`}
