+++ src/components/footer/footer.tsx
@@ -1,0 +1,13 @@
+export const Footer: React.FC = () => {
+  return (
+      <footer className="flex bg-[#2B8AFF] text-[white] p-4 text-center w-screen fixed bottom-[0] left-[0] right-[0] justify-center">
+          <div className="flex justify-between w-[1536px] items-center">
+              <p>
+                  Команда: Дарья Басынина, Алексей Осипенко, Алексей Пономарев,
+                  Дарья Слободская
+              </p>
+              <p>2025</p>
+          </div>
+      </footer>
+  )
+};
