+++ src/components/index.js
@@ -1,6 +1,12 @@
-export * from './button/button'
-export * from './header/header'
+export * from './book-card/book-card'
+export * from './buttons/button'
+export * from './buttons/form-button'
+export * from './filters/filters'
 export * from './icon/icon'
 export * from './icon-button/icon-button'
 export * from './input/input'
+export * from './recently-viewed/recently-viewed'
+export * from './header/header'
+export * from './search/search'
 export * from './validation-error/validation-error'
+export * from './footer/footer'
