+++ src/pages/index.js
@@ -1,1 +1,2 @@
 export * from './authorization/authorization'
+export * from './main-page/main-page'
