+++ src/types/components.d.ts
@@ -1,9 +1,15 @@
 declare module '@components' {
-    export { Header } from '../components/header/header'
-    export { Logo } from '../components/header/components/logo/logo'
+    export { Button } from '../components/buttons/button'
+    export { FormButton } from '../components/buttons/form-button'
+    export { Filters } from '../components/filters/filters'
+    export { BookCard } from '../components/book-card/book-card'
     export { Icon } from '../components/icon/icon'
-    export { Routes } from '../components/constans/routes'
     export { IconButton } from '../components/icon-button/icon-button'
     export { Input } from '../components/input/input'
-    export { Button } from '../components/button/button'
+    export { Footer } from '../components/footer/footer'
+    export { Header } from '../components/header/header'
+    export { Logo } from '../components/header/components/logo/logo'
+    export { RecentlyViewed } from '../components/recently-viewed/recently-viewed'
+    export { Routes } from '../components/constans/routes'
+    export { Search } from '../components/search/search'
 }
