+++ src/types/pages.d.ts
@@ -1,3 +1,4 @@
 declare module '@pages' {
     export { Authorization } from '../pages/authorization/authorization'
+    export { MainPage } from '../pages/main-page/main-page'
 }
